v0.9.7 - Nurse ekardex in Education
The Nurse ekardex in Education project was incubated at the Teaching Nursing Informatics (TNI) course offered by the University of the Philippines - College of Nursing. The group responsible for the project is composed of members: Cristina Rebollido, Deannah Lou Tampus, Jose III Barbudo, Merly Gregorio, Raymund John Ang and Zenas Paloma (Batch: October-December, 2020).

Changelog (2020-10-26): Divided display area into 2 columns. Only upper and lower case letters, space, dash and apostrophe as acceptable characters for the name field. Included name of user, date and time when kardex was generated for validation. Added laboratory and diagnostics field as text area. Added instructions as small display text below input fields. Changed IV access from input field to text area. Changed "Nursing Plan of Care" to "Nursing Plan of Care & Reminders".

Changelog (2020-10-29): Make name and age read-only if name and age fields are not empty. Include user, date and time of last edit.

Changelog (2020-12-05): Include "Nurse Admin" as role. Only allow read functions for nurse admins in the frontend; no edit or print functions. Allow dashboard profile access for Nurse and Nurse Admin roles. Make patient name clickable on search result. Allow Nurse Admin role to access encrypted ciphertext for alert demo purpose.

+++ room.php
@@ -12,18 +12,18 @@ add_action( 'init', 'rja_register_room_cpt' );
 function rja_register_room_cpt()
 {	
 	$labels = array(
-	    'name'				=> _x( 'Rooms', 'Post type general name', 'open-nis' ),
-	    'singular_name'		=> _x( 'Room', 'Post type singular name', 'open-nis' ),
-	    'menu_name'			=> _x( 'Rooms', 'Admin Menu text', 'open-nis' ),
-	    'name_admin_bar'	=> _x( 'Room', 'Add New on Toolbar', 'open-nis' ),
-	    'add_new'			=> __( 'Add New', 'open-nis' ),
-	    'add_new_item'		=> __( 'Add New Room', 'open-nis' ),
-	    'new_item'			=> __( 'New Room', 'open-nis' ),
-	    'edit_item'			=> __( 'Edit Room', 'open-nis' ),
-	    'view_item'			=> __( 'View Room', 'open-nis' ),
-	    'all_items'			=> __( 'All Rooms', 'open-nis' ),
-	    'search_items'		=> __( 'Search Rooms', 'open-nis' ),
-	    'not_found'			=> __( 'No Rooms found.', 'open-nis' )
+	    'name'           => _x( 'Rooms', 'Post type general name', 'open-nis' ),
+	    'singular_name'  => _x( 'Room', 'Post type singular name', 'open-nis' ),
+	    'menu_name'      => _x( 'Rooms', 'Admin Menu text', 'open-nis' ),
+	    'name_admin_bar' => _x( 'Room', 'Add New on Toolbar', 'open-nis' ),
+	    'add_new'        => __( 'Add New', 'open-nis' ),
+	    'add_new_item'   => __( 'Add New Room', 'open-nis' ),
+	    'new_item'       => __( 'New Room', 'open-nis' ),
+	    'edit_item'      => __( 'Edit Room', 'open-nis' ),
+	    'view_item'      => __( 'View Room', 'open-nis' ),
+	    'all_items'      => __( 'All Rooms', 'open-nis' ),
+	    'search_items'   => __( 'Search Rooms', 'open-nis' ),
+	    'not_found'      => __( 'No Rooms found.', 'open-nis' )
 	);
 
 	$args = array(
@@ -34,7 +34,23 @@ function rja_register_room_cpt()
 	    'show_in_menu'       => true,
 	    'query_var'          => true,
 	    'rewrite'            => array( 'slug' => 'room' ),
-	    'capability_type'    => 'post',
+	    'map_meta_cap'       => true,
+	    'capability_type'    => array( 'room', 'rooms' ),
+	    'capabilities'       => array(
+	    	'read_post'                 => 'read_room',
+	    	'read_private_posts'        => 'read_private_rooms',
+	    	'publish_post'              => 'publish_room',
+	    	'publish_posts'             => 'publish_rooms',
+	    	'edit_post'                 => 'edit_room',
+	    	'edit_posts'                => 'edit_rooms',
+	    	'edit_published_posts'      => 'edit_published_rooms',
+	    	'edit_others_post'          => 'edit_others_room',
+	    	'edit_others_posts'         => 'edit_others_rooms',
+	    	'delete_post'               => 'delete_room',
+	    	'delete_posts'              => 'delete_rooms',
+	    	'delete_published_posts'    => 'delete_published_rooms',
+	    	'delete_others_posts'       => 'delete_others_rooms'
+	    ),
 	    'has_archive'        => true,
 	    'hierarchical'       => true,
 	    'menu_position'      => null,
@@ -44,13 +60,13 @@ function rja_register_room_cpt()
 	register_post_type( 'room', $args );
 }
 
-// Single Room - Before Header
+// Single Room - before header
 add_action( 'wp', 'rja_single_room_content_header' );
 
 function rja_single_room_content_header()
 {
 
-	if ( isset($_POST['edit-room']) && wp_verify_nonce($_POST['token'], 'token') ) {	
+	if ( isset($_POST['edit-room']) && wp_verify_nonce($_POST['token'], 'token') && ( current_user_can('administrator') || current_user_can('nurse') ) ) {	
 
 		$room = array(
 			'ID' => $_POST['pid'],
@@ -59,7 +75,11 @@ function rja_single_room_content_header()
 			);	
 
 		$pid = wp_update_post($room);
+		$current_user = wp_get_current_user();
 
+		update_metadata( 'post', $pid, 'room_last_edit_user', Basic::encrypt($current_user->display_name) );
+		update_metadata( 'post', $pid, 'room_last_edit_date', Basic::encrypt(date("Y-m-d")) );
+		update_metadata( 'post', $pid, 'room_last_edit_time', Basic::encrypt(date("h:i:sa")) );
 		update_metadata( 'post', $pid, 'room_name', Basic::encrypt($_POST['room-name']) );
 		update_metadata( 'post', $pid, 'room_age', Basic::encrypt($_POST['age']) );
 		update_metadata( 'post', $pid, 'room_sex', Basic::encrypt($_POST['sex']) );
@@ -74,6 +94,7 @@ function rja_single_room_content_header()
 		update_metadata( 'post', $pid, 'room_bowel', Basic::encrypt($_POST['bowel']) );
 		update_metadata( 'post', $pid, 'room_history', Basic::encrypt($_POST['history']) );
 		update_metadata( 'post', $pid, 'room_medical_notes', Basic::encrypt($_POST['medical-notes']) );
+		update_metadata( 'post', $pid, 'room_diagnostics', Basic::encrypt($_POST['diagnostics']) );
 		update_metadata( 'post', $pid, 'room_nursing_plan', Basic::encrypt($_POST['nursing-plan']) );
 
 		$link = get_permalink($pid);
@@ -81,7 +102,7 @@ function rja_single_room_content_header()
 
 	}
 
-	if ( isset($_POST['delete-room']) && isset($_POST['pid']) ) {
+	if ( isset($_POST['delete-room']) && isset($_POST['pid']) && wp_verify_nonce($_POST['token'], 'token') && current_user_can('administrator') ) {
 
 		wp_delete_post( $_POST['pid'], true );
 		wp_redirect( home_url() );
@@ -130,9 +151,13 @@ function rja_single_room_content()
 
 			$room = get_metadata( 'post', get_the_ID() );
 
+			$room_last_edit_user = Basic::decrypt($room['room_last_edit_user'][0]);
+			$room_last_edit_date = Basic::decrypt($room['room_last_edit_date'][0]);
+			$room_last_edit_time = Basic::decrypt($room['room_last_edit_time'][0]);
 			$room_name = Basic::decrypt($room['room_name'][0]);
 			$room_age = Basic::decrypt($room['room_age'][0]);
-			if ( Basic::decrypt($room['room_sex'][0]) == 'M' ) { $room_sex = 'Male'; } else { $room_sex = 'Female'; }
+			if ( Basic::decrypt($room['room_sex'][0]) == 'M' ) { $room_sex = 'Male'; }
+			if ( Basic::decrypt($room['room_sex'][0]) == 'F' ) { $room_sex = 'Female'; }
 			$room_date_admission = Basic::decrypt($room['room_date_admission'][0]);
 			$room_doctor = Basic::decrypt($room['room_doctor'][0]);
 			$room_reason = Basic::decrypt($room['room_reason'][0]);
@@ -144,26 +169,35 @@ function rja_single_room_content()
 			$room_bowel = Basic::decrypt($room['room_bowel'][0]);
 			$room_history = Basic::decrypt($room['room_history'][0]);
 			$room_medical_notes = Basic::decrypt($room['room_medical_notes'][0]);
+			$room_diagnostics = Basic::decrypt($room['room_diagnostics'][0]);
 			$room_nursing_plan = Basic::decrypt($room['room_nursing_plan'][0]);
 
 			?>
 			<div id="kardex">
-				<h3>Room: <?php esc_html(the_title()); ?></h3>
-				<p>Name: <?= esc_html($room_name); ?><br />
-				Age: <?= esc_html($room_age); ?><br />
-				Sex: <?= esc_html($room_sex); ?></p>
-				<p>Admission Date: <?= esc_html($room_date_admission); ?><br />
-				Doctor: <?= esc_html($room_doctor); ?><br />
-				Reason: <?= esc_html($room_reason); ?><br />
-				Allergy: <?= esc_html($room_allergy); ?><br />
-				Diet: <?= esc_html($room_diet); ?></p>
-				<p>IV Access: <?= esc_html($room_iv_access); ?><br />
-				Monitoring: <?= esc_html($room_monitoring); ?><br />
-				Urine: <?= esc_html($room_urine); ?><br />
-				Bowel Movement: <?= esc_html($room_bowel); ?></p>
-				<p><em>History:</em><br /><?= nl2br(esc_html($room_history)); ?></p>
-				<p><em>Medical Notes:</em><br /><?= nl2br(esc_html($room_medical_notes)); ?></p>
-				<p><em>Nursing Plan of Care:</em><br /><?= nl2br(esc_html($room_nursing_plan)); ?></p>
+				<div style="float: left; width: 220px; margin-right: 120px;">
+					<h3>Room: <?php esc_html(the_title()); ?></h3>
+					<p>Name: <?= esc_html($room_name); ?><br />
+					Age: <?= esc_html($room_age); ?>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Sex: <?= esc_html($room_sex); ?></p>
+					<p>Admission Date: <?= esc_html($room_date_admission); ?><br />
+					Doctor: <?= esc_html($room_doctor); ?><br />
+					Reason: <?= esc_html($room_reason); ?><br />
+					Allergy: <?= esc_html($room_allergy); ?><br />
+					Diet: <?= esc_html($room_diet); ?></p>
+					<p>IV Access:<br />
+					<?= nl2br(esc_html($room_iv_access)); ?><br />
+					<br />
+					Monitoring: <?= esc_html($room_monitoring); ?><br />
+					Urine: <?= esc_html($room_urine); ?><br />
+					Bowel Movement: <?= esc_html($room_bowel); ?></p>
+				</div>
+				<div style="float: left; width: 340px;">
+					<p><em>History:</em><br /><?= nl2br(esc_html($room_history)); ?></p>
+					<p><em>Medical Notes:</em><br /><?= nl2br(esc_html($room_medical_notes)); ?></p>
+					<p><em>Laboratory & Diagnostics:</em><br /><?= nl2br(esc_html($room_diagnostics)); ?></p>
+					<p><em>Nursing Plan of Care & Reminders:</em><br /><?= nl2br(esc_html($room_nursing_plan)); ?></p>
+				</div>
+				<small style="display: block; clear: both;">Kardex was last edited by <?= $room_last_edit_user ?> on <?= $room_last_edit_date ?> at <?= $room_last_edit_time ?>.</small>
+				<small style="display: block; clear: both;">Kardex report generated by <?php $current_user = wp_get_current_user(); echo $current_user->display_name; ?> on <?php echo date("Y-m-d"); ?> at <?php echo date("h:i:sa"); ?>.</small>
 				<div id="no-print">
 					<h2>Edit Information</h2>
 					<button onclick="toggleRoomForm()">Edit Form</button>
@@ -178,11 +212,11 @@ function rja_single_room_content()
 							</p>
 							<p>
 								<label for="name">Name</label><br />
-								<input type="text" id="room-name" name="room-name" value="<?= esc_html($room_name); ?>" required />
+								<input type="text" id="room-name" name="room-name" value="<?= esc_html($room_name); ?>" pattern="[A-Za-z '-]+" title="Patient name" <?php echo ( ($room_name == '') ? 'required' : 'readonly' ); ?> />
 							</p>
 							<p>
 								<label for="age">Age</label><br />
-								<input type="number" id="age" name="age" value="<?= esc_html($room_age); ?>" required />
+								<input type="number" id="age" name="age" value="<?= esc_html($room_age); ?>" <?php echo ( ($room_age == '') ? 'required' : 'readonly' ); ?> />
 							</p>
 							<p>
 								<label for="sex">Sex</label><br />
@@ -209,23 +243,28 @@ function rja_single_room_content()
 							</p>
 							<p>
 								<label for="diet">Diet</label><br />
-								<input type="text" id="diet" name="diet" value="<?= esc_html($room_diet);?>" required /><br />
+								<input type="text" id="diet" name="diet" value="<?= esc_html($room_diet);?>" required />
+								<small>Standing diet order and special instructions.</small>
 							</p>
 							<p>
 								<label for="iv-access">IV Access</label><br />
-								<input type="text" id="iv-access" name="iv-access" value="<?= esc_html($room_iv_access);?>" required /><br />
+								<textarea id="iv-access" name="iv-access" required><?= esc_html($room_iv_access); ?></textarea>
+								<small>Central and/or peripheral. Size, location, fluids or locked. New line each site.</small>
 							</p>
 							<p>
 								<label for="monitoring">Monitoring</label><br />
-								<input type="text" id="monitoring" name="monitoring" value="<?= esc_html($room_monitoring);?>" required /><br />
+								<input type="text" id="monitoring" name="monitoring" value="<?= esc_html($room_monitoring);?>" required />
+								<small>Vital signs, neurological, vascular checks, etc. and frequency.</small>
 							</p>
 							<p>
 								<label for="urine">Urine</label><br />
-								<input type="text" id="urine" name="urine" value="<?= esc_html($room_urine);?>" required /><br />
+								<input type="text" id="urine" name="urine" value="<?= esc_html($room_urine);?>" required />
+								<small>Description (i.e. color, transparency).</small>
 							</p>
 							<p>
 								<label for="bowel">Bowel Movement</label><br />
-								<input type="text" id="bowel" name="bowel" value="<?= esc_html($room_bowel);?>" required /><br />
+								<input type="text" id="bowel" name="bowel" value="<?= esc_html($room_bowel);?>" required />
+								<small>Last BM, description.</small>
 							</p>
 							<p>
 								<label for="history">History</label><br />
@@ -234,25 +273,93 @@ function rja_single_room_content()
 							<p>
 								<label for="medical-notes">Medical Notes</label><br />
 								<textarea id="medical-notes" name="medical-notes" required><?= esc_html($room_medical_notes); ?></textarea>
+								<small>Pertinent medical notes and plan of care.</small>
 							</p>
 							<p>
-								<label for="nursing-plan">Nursing Plan of Care</label><br />
+								<label for="diagnostics">Laboratory and Diagnostics</label><br />
+								<textarea id="diagnostics" name="diagnostics" required><?= esc_html($room_diagnostics); ?></textarea>
+								<small>Blood works, imaging and other diagnostics. Indicate if pending or done.</small>
+							</p>
+							<p>
+								<label for="nursing-plan">Nursing Plan of Care & Reminders</label><br />
 								<textarea id="nursing-plan" name="nursing-plan" required><?= esc_html($room_nursing_plan); ?></textarea>
+								<small>Nursing notes, plan of care & reminders.</small>
 							</p>
 							<p>
 								<input type="submit" value="Edit Patient" id="edit-room" name="edit-room" />
 							</p>
 							<!-- <p><input type="reset" value="Clear Details" id="reset-form" name="reset-form" /></p> -->
-							<?php if (current_user_can('administrator')) : ?>
-								<p><input type="submit" value="Delete Room" id="delete-room" name="delete-room" onclick="return confirm('Are you sure you want to delete this room?');" /></p>
-							<?php endif ?>
 							<?php wp_nonce_field( 'token', 'token' ); ?>
 						</form>
+							<?php if (current_user_can('administrator')) : ?>
+								<form method="post">
+									<p>
+										<input type="submit" value="Delete Room" id="delete-room" name="delete-room" onclick="return confirm('Are you sure you want to delete this room?');" />
+									</p>
+									<input type="hidden" id="pid" name="pid" value="<?php the_ID(); ?>" readonly />
+									<?php wp_nonce_field( 'token', 'token' ); ?>
+								</form>
+							<?php endif ?>
 						<button onclick="toggleRoomForm()">Close Form</button>
 					</div>
 				</div>
 			</div>
 			<?php			
+		} elseif ( current_user_can('nurse_admin') ) {
+
+			$room = get_metadata( 'post', get_the_ID() );
+
+			$room_last_edit_user = Basic::decrypt($room['room_last_edit_user'][0]);
+			$room_last_edit_date = Basic::decrypt($room['room_last_edit_date'][0]);
+			$room_last_edit_time = Basic::decrypt($room['room_last_edit_time'][0]);
+			$room_name = Basic::decrypt($room['room_name'][0]);
+			$room_age = Basic::decrypt($room['room_age'][0]);
+			if ( Basic::decrypt($room['room_sex'][0]) == 'M' ) { $room_sex = 'Male'; }
+			if ( Basic::decrypt($room['room_sex'][0]) == 'F' ) { $room_sex = 'Female'; }
+			$room_date_admission = Basic::decrypt($room['room_date_admission'][0]);
+			$room_doctor = Basic::decrypt($room['room_doctor'][0]);
+			$room_reason = Basic::decrypt($room['room_reason'][0]);
+			$room_allergy = Basic::decrypt($room['room_allergy'][0]);
+			$room_diet = Basic::decrypt($room['room_diet'][0]);
+			$room_iv_access = Basic::decrypt($room['room_iv_access'][0]);
+			$room_monitoring = Basic::decrypt($room['room_monitoring'][0]);
+			$room_urine = Basic::decrypt($room['room_urine'][0]);
+			$room_bowel = Basic::decrypt($room['room_bowel'][0]);
+			$room_history = Basic::decrypt($room['room_history'][0]);
+			$room_medical_notes = Basic::decrypt($room['room_medical_notes'][0]);
+			$room_diagnostics = Basic::decrypt($room['room_diagnostics'][0]);
+			$room_nursing_plan = Basic::decrypt($room['room_nursing_plan'][0]);
+			
+			?>
+			<div id="kardex">
+				<div style="float: left; width: 220px; margin-right: 120px;">
+					<h3>Room: <?php esc_html(the_title()); ?></h3>
+					<p>Name: <?= esc_html($room_name); ?><br />
+					Age: <?= esc_html($room_age); ?>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Sex: <?= esc_html($room_sex); ?></p>
+					<p>Admission Date: <?= esc_html($room_date_admission); ?><br />
+					Doctor: <?= esc_html($room_doctor); ?><br />
+					Reason: <?= esc_html($room_reason); ?><br />
+					Allergy: <?= esc_html($room_allergy); ?><br />
+					Diet: <?= esc_html($room_diet); ?></p>
+					<p>IV Access:<br />
+					<?= nl2br(esc_html($room_iv_access)); ?><br />
+					<br />
+					Monitoring: <?= esc_html($room_monitoring); ?><br />
+					Urine: <?= esc_html($room_urine); ?><br />
+					Bowel Movement: <?= esc_html($room_bowel); ?></p>
+				</div>
+				<div style="float: left; width: 340px;">
+					<p><em>History:</em><br /><?= nl2br(esc_html($room_history)); ?></p>
+					<p><em>Medical Notes:</em><br /><?= nl2br(esc_html($room_medical_notes)); ?></p>
+					<p><em>Laboratory & Diagnostics:</em><br /><?= nl2br(esc_html($room_diagnostics)); ?></p>
+					<p><em>Nursing Plan of Care & Reminders:</em><br /><?= nl2br(esc_html($room_nursing_plan)); ?></p>
+				</div>
+				<small style="display: block; clear: both;">Kardex was last edited by <?= $room_last_edit_user ?> on <?= $room_last_edit_date ?> at <?= $room_last_edit_time ?>.</small>
+				<small style="display: block; clear: both;">Kardex report generated by <?php $current_user = wp_get_current_user(); echo $current_user->display_name; ?> on <?php echo date("Y-m-d"); ?> at <?php echo date("h:i:sa"); ?>.</small>
+				<div id="no-print">
+				</div>
+			</div>
+			<?php
 		} else {
 			?>
 			<p>You do not have permission to view patient information.</p>
